[REFACTOR][IR] Inline ApplyPassToFunction into relax decompose_ops, delete the standalone util
ApplyPassToFunction was a general-purpose "run this pass on only the
function whose name matches a regex" wrapper, but its sole in-tree
production callers are DecomposeOpsForInference / DecomposeOpsForTraining
in src/relax/transform/decompose_ops.cc.  Inline the logic as a
file-local helper there (using exact-name match rather than a regex,
since all callers supply a literal function name), delete
src/ir/apply_pass_to_function.cc, remove its transform.h declaration,
and remove the Python wrapper.  Also remove the two
DCE-with-ApplyPassToFunction tests in
test_transform_dead_code_elimination.py — they exercise the util's
plumbing rather than DCE behavior.
